220831 compability

+++ PKHeX_Raid_Plugin/Raid/RaidPKM.cs
@@ -3,12 +3,12 @@
     public class RaidPKM
     {
         public readonly int Nature;
-        public readonly int Species;
+        public readonly ushort Species;
         public readonly int[] IVs;
         public readonly int Gender;
         public readonly uint EC;
         public readonly uint PID;
-        public readonly int AltForm;
+        public readonly byte AltForm;
         public readonly int Ability;
         public readonly uint ShinyType;
         public readonly bool IsGigantamax;
@@ -16,7 +16,7 @@
         public readonly int Deviation;
         public readonly int AbilityIdx;
 
-        public RaidPKM(int species, int altForm, uint ec, uint pid, int[] ivs, int ability, int abilityIdx, int gender, int nature, int deviation, uint shinyType, bool isGigantamax, sbyte forcedShinyType = 0)
+        public RaidPKM(ushort species, byte altForm, uint ec, uint pid, int[] ivs, int ability, int abilityIdx, int gender, int nature, int deviation, uint shinyType, bool isGigantamax, sbyte forcedShinyType = 0)
         {
             Species = species;
             AltForm = altForm;
